refactor(poll-modal): move to poll folder
