+++ src/screens/poll-screen/modals/cant-create-poll/styles.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components/native';
 import Colors from '../../../../constants/colors';
-import button from '../../../button';
+import button from '../../../../components/button';
 
 const Container = styled.View`
   display: flex;
